Factor MSBuildLocator test logic into base class
